pyproject.toml
[project]
name = "hive-mcp-gateway"
version = "0.3.0"
description = "Intelligent MCP gateway and tool management system - part of The Hive agentic coding ecosystem"
readme = "README.md"
authors = [
    { name = "KHAEntertainment", email = "contact@khaentertainment.com" }
]
repository = "https://github.com/KHAEntertainment/hive-mcp-gateway"
homepage = "https://github.com/KHAEntertainment/hive-mcp-gateway"
keywords = ["mcp", "ai", "tools", "gateway", "hive", "agentic", "proxy"]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "fastapi-mcp",
    "numpy>=1.24.0",
    "sentence-transformers>=2.2.0",
    "aiohttp>=3.9.0",
    "aiofiles>=23.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.25.0",
    "rich>=13.0.0",
    "mcp>=0.1.0",
    "watchfiles>=0.21.0",
    "PyQt6>=6.6.0",
    "PyQt6-WebEngine>=6.6.0",
    "psutil>=5.9.0",
    "keyring>=24.0.0",
    "oauthlib>=3.2.0",
    "requests-oauthlib>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
hive-mcp-gateway = "hive_mcp_gateway:main"
hive-mcp = "hive_mcp_gateway:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "UP"]
ignore = []

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.uv.sources]
fastapi-mcp = { git = "https://github.com/tadata-org/fastapi_mcp.git" }

[dependency-groups]
dev = [
    "black>=25.1.0",
    "httpx>=0.28.1",
    "mypy>=1.15.0",
    "pyinstaller>=6.0.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "ruff>=0.11.11",
]
